+++ py/dynesty/bounding.py
@@ -42,7 +42,7 @@ __all__ = [
 ]
 
 BOUND_LIST = ['none', 'single', 'multi', 'balls', 'cubes']
-# these are the exitsing bounds implemented here
+# these are the existing bounds implemented here
 
 
 def _slogdet_checked(matrix):
@@ -1214,6 +1214,7 @@ class SupFriends(Bound):
         self.am /= hsmax**2
         self.axes *= hsmax
         self.axes_inv /= hsmax
+        self.ctrs = points
 
         detln = _slogdet_checked(self.am)
         self.logvol = self.ndim * np.log(2.) - 0.5 * detln
@@ -1221,8 +1222,7 @@ class SupFriends(Bound):
         # Estimate the volume and fractional overlap with the unit cube
         # using Monte Carlo integration.
         if mc_integrate:
-            self.funit = self.monte_carlo_logvol(points,
-                                                 return_overlap=True,
+            self.funit = self.monte_carlo_logvol(return_overlap=True,
                                                  rstate=rstate)[1]
 
     def _get_covariance_from_all_points(self, points):